// src/Server/Infrastructure/PgTyped/Schemas/__generated__/users.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: ucC2qx40t5aeB2Jx7K/TbUCISXCuvOdJKg8rc6pbHAMAYAQ5j3aw1XidE2fNN5/IG/bX7KTbEJGj7bbEzo/wzw==
 */

/* eslint-disable */
// tslint:disable

import Cities from './cities'
import Countries from './countries'
import States from './states'

interface Users {
  address_complement: (string) | null
  address_number: (string) | null
  address_street: (string) | null
  city_id: Cities['id']
  country_id: Countries['id']
  /**
   * @default '2022-06-27'::date
   */
  created_at: Date
  document_number: string
  email: string
  first_name: string
  /**
   * @default nextval('users_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'users_id'}
  /**
   * @default false
   */
  is_active: (boolean) | null
  last_name: string
  middle_name: (string) | null
  neighbourhood: (string) | null
  postal_code: (string) | null
  state_id: States['id']
  /**
   * @default '2022-06-27'::date
   */
  updated_at: Date
  /**
   * @default false
   */
  was_activated: (boolean) | null
  whatsapp_id: string
  whatsapp_number: string
}
export default Users;

interface Users_InsertParameters {
  address_complement?: (string) | null
  address_number?: (string) | null
  address_street?: (string) | null
  city_id: Cities['id']
  country_id: Countries['id']
  /**
   * @default '2022-06-27'::date
   */
  created_at?: Date
  document_number: string
  email: string
  first_name: string
  /**
   * @default nextval('users_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'users_id'}
  /**
   * @default false
   */
  is_active?: (boolean) | null
  last_name: string
  middle_name?: (string) | null
  neighbourhood?: (string) | null
  postal_code?: (string) | null
  state_id: States['id']
  /**
   * @default '2022-06-27'::date
   */
  updated_at?: Date
  /**
   * @default false
   */
  was_activated?: (boolean) | null
  whatsapp_id: string
  whatsapp_number: string
}
export type {Users_InsertParameters}